+++ yall/python/yall/networktables/targets.py
@@ -1,7 +1,7 @@
 from dataclasses import dataclass, field
 from typing import Any, Dict, List, Optional
 
-from util import LimelightUtils
+from .util import LimelightUtils
 from wpimath import geometry
